requirements.txt
absl-py==0.11.0
alembic==1.5.5
amqp==5.0.6
antlr4-python3-runtime==4.8
appdirs==1.4.4
Authlib==0.15.3
awscli==1.19.84
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
black==21.4b2
bleach==3.3.0
blinker==1.4
blis==0.7.4
boto==2.49.0
boto3==1.17.84
botocore==1.20.84
Bottleneck==1.3.2
cached-property==1.5.2
cachetools==4.2.1
catalogue==1.0.0
celery==5.1.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cloudpickle==1.6.0
colorama==0.4.3
cryptography==3.4.6
cycler==0.10.0
cymem==2.0.5
Cython==0.29.22
decorator==4.4.2
dnspython==2.1.0
docutils==0.15.2
email-validator==1.1.2
et-xmlfile==1.0.1
fastai==2.4
fastcore==1.3.19
fastprogress==1.0.0
Flask==2.0.1
Flask-Admin==1.5.7
Flask-Babel
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
fpdf2==2.4.1
future==0.18.2
fvcore==0.1.5.post20210630
google-api-core==1.26.0
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-cloud-vision==2.4.2
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.36.0
gunicorn==20.1.0
hydra-core==1.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.7.0
importlib-resources==5.2.0
imutils==0.5.4
iopath==0.1.8
itsdangerous==2.0.1
jdcal==1.4.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
kombu==5.1.0
libcst==0.3.17
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mock==4.0.3
murmurhash==1.0.5
mypy-extensions==0.4.3
mysqlclient==2.0.3
networkx==2.5
numexpr==2.7.2
numpy==1.20.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
omegaconf==2.1.0
opencv-python==4.5.1.48
openpyxl==3.0.6
packaging==20.9
pandas==1.2.2
pathspec==0.8.1
pathy==0.4.0
pdf2image==1.14.0
Pillow==8.2.0
plac==1.1.3
portalocker==2.2.1
preshed==3.0.5
prompt-toolkit==3.0.18
proto-plus==1.19.0
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-Levenshtein==0.12.0
python-slugify==4.0.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
regex==2021.7.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.4.2
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
smart-open==3.0.0
soupsieve==2.2
spacy==2.3.5
spacy-legacy==3.0.1
SQLAlchemy==1.3.23
srsly==1.0.5
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
text-unidecode==1.3
thinc==7.4.5
threadpoolctl==2.1.0
tifffile==2021.2.26
toml==0.10.2
torch==1.8.0
torchvision==0.9.0
tqdm==4.58.0
typed-ast==1.4.3
typer==0.3.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
vine==5.0.0
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
WTForms==2.3.3
XlsxWriter==1.3.7
yacs==0.1.8
zipp==3.4.0